[enh] fetch: added optimization for blog.hu feeds
blog.hu feeds don't support Last-Modified nor ETag, so in order to avoid
parsing the RSS when nothing changed, special treatment has been added.
It strips the sometimes-chaning <generator>...</generator> tag and
calculates a SHA-1 hash of the result. If it remains unchanged, it
treats the feed the same way as an HTTP 304 response.