[enh] optimized archive implementation
The current implementation loads every field of the `Feed` item, and
then causes an update to every field (including the content!). This
commit changes this to a single update, and the return value containing
the rows matched can still be used to determine the success.

See https://docs.djangoproject.com/en/dev/ref/models/querysets/#update